+++ lib/nsrr/helpers/download_request.rb
@@ -1,9 +1,12 @@
+# frozen_string_literal: true
+
 require 'openssl'
 require 'net/http'
 require 'uri'
 
 module Nsrr
   module Helpers
+    # Downloads a file to a specified folder.
     class DownloadRequest
       class << self
         def get(*args)
@@ -14,19 +17,17 @@ module Nsrr
       attr_reader :url, :error, :file_size
 
       def initialize(url, download_folder)
-        begin
-          escaped_url = URI.escape(url)
-          @url = URI.parse(escaped_url)
-          @http = Net::HTTP.new(@url.host, @url.port)
-          if @url.scheme == 'https'
-            @http.use_ssl = true
-            @http.verify_mode = OpenSSL::SSL::VERIFY_NONE
-          end
-          @download_folder = download_folder
-          @file_size = 0
-        rescue => e
-          @error = 'Invalid Token'
+        escaped_url = URI.escape(url)
+        @url = URI.parse(escaped_url)
+        @http = Net::HTTP.new(@url.host, @url.port)
+        if @url.scheme == 'https'
+          @http.use_ssl = true
+          @http.verify_mode = OpenSSL::SSL::VERIFY_NONE
         end
+        @download_folder = download_folder
+        @file_size = 0
+      rescue
+        @error = 'Invalid Token'
       end
 
       # Writes file segments to disk immediately instead of storing in memory
@@ -35,7 +36,8 @@ module Nsrr
         begin
           partial = true
           @http.request_get(@url.path) do |response|
-            case response.code when '200'
+            case response.code
+            when '200'
               response.read_body do |segment|
                 local_file.write(segment)
               end
@@ -50,13 +52,10 @@ module Nsrr
         rescue => e # Net::ReadTimeout, SocketError
           @error = "(#{e.class}) #{e.message}"
         ensure
-          local_file.close()
-          ::File.delete(@download_folder) if partial and ::File.exist?(@download_folder)
+          local_file.close
+          ::File.delete(@download_folder) if partial && ::File.exist?(@download_folder)
         end
       end
-
     end
   end
 end
-
-